backend/requirements.txt
earthengine-api
python-dotenv
google-generativeai
openai
folium
fastapi
pydantic
uvicorn
